+++ compiler-tests/src/test/data/box/inject/assisted/GenericAssistedParamsAcrossModules.kt
@@ -1,0 +1,64 @@
+// MODULE: lib
+@Inject
+class Example<T>(@Assisted val inputT: T, val graphT: T) {
+  @AssistedFactory
+  fun interface Factory<T> {
+    fun create(input: T): Example<T>
+  }
+
+  @AssistedFactory
+  fun interface Factory2 {
+    fun create(input: Int): Example<Int>
+  }
+}
+
+@Inject
+class ExampleWithDifferent<T, R>(@Assisted val inputT: T, val graphT: R) {
+  @AssistedFactory
+  fun interface Factory<T, R> {
+    fun create(input: T): ExampleWithDifferent<T, R>
+  }
+
+  @AssistedFactory
+  fun interface Factory2<T> {
+    fun create(input: Int): ExampleWithDifferent<Int, T>
+  }
+}
+
+// MODULE: main(lib)
+@DependencyGraph
+interface AppGraph {
+  val exampleFactory: Example.Factory<Int>
+  val exampleFactory2: Example.Factory2
+
+  val exampleFactory3: ExampleWithDifferent.Factory<Int, Int>
+  val exampleFactory4: ExampleWithDifferent.Factory2<Int>
+
+  @Provides
+  val int: Int
+    get() = 2
+}
+
+fun box(): String {
+  val graph = createGraph<AppGraph>()
+
+  val factory = graph.exampleFactory
+  val example = factory.create(3)
+  assertEquals(3, example.inputT)
+  assertEquals(2, example.graphT)
+  val factory2 = graph.exampleFactory2
+  val example2 = factory.create(3)
+  assertEquals(3, example2.inputT)
+  assertEquals(2, example2.graphT)
+
+  val factory3 = graph.exampleFactory3
+  val example3 = factory3.create(3)
+  assertEquals(3, example3.inputT)
+  assertEquals(2, example3.graphT)
+  val factory4 = graph.exampleFactory4
+  val example4 = factory.create(3)
+  assertEquals(3, example4.inputT)
+  assertEquals(2, example4.graphT)
+
+  return "OK"
+}